+++ src/nF_EBD_Controller.hpp
@@ -8,8 +8,8 @@ class nF_EBD_Controller {
  public:
   nF_EBD_Controller(const char *com_name);
   void init();
-  std::array<float, 2> read();
-  void move_to(std::array<float, 2> target);
+  std::array<double, 2> read();
+  void move_to(std::array<double, 2> target);
   void close();
 
  private:
@@ -18,7 +18,7 @@ class nF_EBD_Controller {
   int fd;                       // file descriptor, to send commands to the stage
   int axis0 = 0;                // x-axis
   int axis1 = 1;                // y-axis
-  float offset = 1000.0;        // urad
+  double offset = 1000.0;       // urad
   std::atomic<bool> is_moving;  // stage is unreachable while moving
-  void move_to_blocking(float x_target, float y_target);
+  void move_to_blocking(double x_target, double y_target);
 };
